routes: move routes to a separate file
